+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Fixed that purpose should be passed on when using Locator.locate_many_signed(strings, for: 'purpose').
+
+    *DHH*
+
 *   Added #to_gid_param and #to_sgid_param as short-cuts for to_gid.to_param and to_sgid.to_param for convenience.
 
     *DHH*